Rework the --help and --version messages
Move the copyright attribution to the --version message so that the
--help message fits on one 80x25 console screen.

Add the commit hash to the --version message if cabextract is compiled
from a git source.

Print the --help message to stdout instead of stderr so that it can be
MOREd on the DOS console. More generally, if a user asks for --help,
then the output should be pageable without needing the 2>&1 redirector,
and the exit status should be success.